Tidy up and amend process_csv_to_fhir tests
